Everything in codec.py
